New Responsive News Admin
Experimental subject to further nitpick.

The reason I did not changed form_select default width to 100% is
because at most of the times, it is loaded at the a 100% wide frame.
But if it is loaded to a right side stacking column div, we can always
use ‘width’=>’100%’. IMO, this gives better control on Dev End, and
much more convenience.