Breadcrumbs Component
The parameter is very good, and fluid.

I tried with the ex-model but the control was not that smooth since the
callback was done systematically on each file. If bound by
generate_breadcrumb(), I couldn’t do much with the template/design end.

That aside, we do not know how they will nest it. In order to grant
maximum flexibility as a component, I propose remove data population
and leave it on template end. Lets just do it all in the template
directly with add_to_breadcrumb(); ?