Multi language the Category Fields.
Time to play with some serialisation. I want to bind an array into a
single column, and call back into multiple fields. When POST, it should
re-validate, rebuild it. Defender wasn’t setup to deal with this, and I
think manual check for required will be necessary here - the primary
one must not be empty. I’ll git later.

With the functions built now, it will be swift to implement on other -
modules, dynamics fields. Tested with category, and work as it should.

Instead of

$data[‘xxx’] = if (isset($_POST[‘xx’])) ? form_sanitizer($_POST[‘xxx’],
‘’, ‘xxx’) : $data[‘xxx’];

Just do:

$data[‘xxx’] = self::fusion_getlocale($data, ‘xxx’);

Consider my version of gettext() puppy.