Since it have been sneaking in, Standard updated, WIP...
Since I have seen more and more Devs not use the ?> closing tag I
figured that there must be a good reason.
And I think that this is one standard we should adapt, hands down.
We are in deep need of a common standard document. So I have started the
work.
Our standards will be a mixture of what we always have done and what
others do if it seems reasonable for now.
Please remember  that the general Standards are WIP,
this change reflects this part.

https://www.php-fusion.co.uk/infusions/wiki/documentation.php?page=209#225