News  - Model - View - Controller, Model Controller View, Model View Controller, Controller Model View The Git is about reinterpreting a theorem in a very literal, and practical method when it's far less much as reinterpreting a non literal but practical code through Fusion's very own conventional procedural approach.
And unless it's not spelled as Model and not wrapped in as an object but does the same thing is not considered a model?
Those want to know more about MVC can have a look at :
http://www.sitepoint.com/the-mvc-pattern-and-php-1/
http://www.sitepoint.com/the-mvc-pattern-and-php-2/
https://r.je/mvc-in-php.html

MVC is a theorem. It's not a coding standard, and if is, therefore failed miserably PHP-Fusion coding standards. That is why, to be reverted.