fixed #201
In a case the sent charset in the connection string was ignored by the MySQL server (using PHP 5.4.30). I could not reproduce this bug on other server even if I tried to set  everything as I saw on the problematic server.

This commit fixes the mysterious bug which is probably related to settings of the server or the PHP version.