Test build prototype 2.0 - #468 The following 3 functions sniffPatterns(); , serializeIdCache(); fetchData(); are rearranged under getOtherTags() - these 3 work together in a single procedural flow and do not need to be segregated.
$this->replacePatterns(); $this->replaceAlias(); are now deprecated. The first part of the code ares removed to prepareStatement() function by reading uniqueID and results from getOtherTags().

Added a new replacement function, which reads only successful $regex_statements and replaced in $ouput.
Added a new debug statement - print_p($this->regex_statements).
Same $regex_statements will show you what statements are like - the search paired with the replacement value in a single array
It will show you the failed pairs that were not replaced which I plan to show in the showQueries function when I manage to hit a milestone.