Massive Lithuanian update
Updated Lithuanian language files, rechecked all files with English ones
and added/deleted what was missing. Please in the future when you do
updates to local files, pleaase update all other languages at the same
time as English, because it takes hell of a time to find what is missing
and not added. Especially if these are small things, like <br /> tags
change to \n and so on.