Removal of pNotify Library - replaced with $defender->stop($notice='') No need to keepup with external notification when we can do it on our own. Those who wanted the notify to pop up, amend your alert() function.
The purpose for this repository to be used was because it was Jquery based, but I couldn't get it to work since it buffers out in the bottom. Not even echo on nested function script triggered it. I'm not willing to mod extensively into the js, and hence, I'm dropping the whole thing since I included it in the first place.

To those who wants to use it still, please do so on your theme level.