Fundation to the Fusion GateWay has been set,
This is a new experimental concept I started in order to make it really hard for bots to be bots !
We use all saves in sessions, dynamic honeypot,, encryption and output in java. same as for example email but harder due to the encryption.
Currently blocking about 30 attempts every 48h on main.