Fix defender filterPostArray function. It must return string instead of array.
The fix is now applicable to the following:
$post_val = (string) \Defender::getInstance()->filterPostArray( array('a', 'b', 'c') ); (Better code)
$post_val = (string) stripinput($_POST['a']['b']['c']); (Less better code)

which is used in form_sanitizer($post_val, '', 'a[b][c]');

Previously this function returns an array, but it should be returning a string instead.

P.s: I know there is another delimiter auto parsing where form_sanitizer() returns 'value1,value2,value3' when multiple options is used. Will test again soon with that scenario soon.