Fix and optimize \Token::generate_token() Previously, when we post to a remote file in the same server, $options['remote_url'] was used to mutate $_SESSION for csrf_tokens and fields. If the calling page could not find the session, all will return blank, and token error will be fired, signalling \Defender to declare FUSION_NULL and therefore safeguard more important parts of the script from intrusion. However, in most cases, the pathing became a problem, where some request could have request_uri (i.e. ?request=key&amp;viewpage=true), or (/folder?q=1). Due to server fault at times, pathing could be be presented like (//folder?q=1) instead. All these has led to problematic implementation, especially where the post file is targetted through a remote ajax script.
As such, it is in my opinion, to replace the output of \Defender::pageHash(), which services the $options['remote_url']. In this patch, \Defender::pageHash() will no longer accept param, and output will always be encrypted. $options['remote_url'] has been deprecated, in favor of, a more stable and permanent md5(SECRET_KEY) - This still essentially gives you a unique $_SESSION hash for the csrf tokens across the board but presented in a more accessible and much more stable for developer to use.

This update is compatible with Andromeda.