#2098 - Lollipop Captcha.
Why Lollipop? I'm currently eating a Lollipop given to me by my daughter, hence the name.
The request idea from @Falk was like this:

Make a randomized output , maybe even a algo..
No, make it an algo..
For example 9 inputs options, randomize them and output 3 possible ones everytime with a twist. Some are pre-checked others are not. Just store correct array and compare for post to success.
That way the bots need to guess 3 randomized check or no check every single post.. I see it virtually impossible before spam filter kick in.

I commented the lines of how I code it, except that for some change.
"3 possible choice", is changed to a "random guess", so bots will never know how many is the correct one. Sometimes, it's 1, sometimes it all 9 options.
"None are checked" at default for better UX, to avoid 'unchecking actions'. It may lead to mistakes when we tend to rush over forms.
"Storage of correct values" will not be done, we'll use validation technique instead. I will outline in the class how to get correct values later after this commit.