Init Commit Inspire AT
- Removes Genesis AT and Merge into Inspire AT.
- Admin Dashboard Widget init.
- Note about Plugins Src and Image Lib. These files are obtained from Inspinia AT for REFERENCES only. I picked it out randomly with the themes available on web for a startup discussions.

PHPFusion will not fork or mod on Inspinia AT, but it is to give an idea to developers the layouts and app views generally that is mainstream to learn and update our knowledge, know-how, and perception, including JS implementations solely for discussions purposes generally on Discord Developer Channels. The test pages is just a reference to use the open-source JS plugins. It's faster to refer to them rather than looking at documentations. With the implementation, we can directly look into implementation in core upgrades more efficiently.

Again, the files are not for publication but development purposes. So don't accuse anyone for copyright infringements BS. We take this very seriously and we do not intend to fork or mod on it.