Adding a security policy
I recommned that PHP-Fusion adds this policy for now so that all the vulnerability reports that are being reported publicly on Github are moved to be private disclosures to protect PHP-Fusion and all the people that use the software. If down the road PHP-Fusion makes a new infusion for the main site to report vulnerabilities then this policy can be updated to direct people there.