Fixed adding user_reputation during infusing forum.
During loading of infusion.php, the column user_reputation does not exist, and therefore if you process the check when it was loaded, it will always return FALSE. The rule will always not loaded. We do not need to even check it. If you left it unchecked, you will get error during installation.

How do we update a non existing column? Place the dbupdatedrow rule when infusion being loaded into the very last key to load in the whole $inf array. It will fix the issue because table will have been modified already.